requirements.txt
Werkzeug==0.16.1
flask
flask-restplus
requests
Flask-SQLAlchemy
PyJWT